Don't test en-us-mathtext.yaml because that is not the table we are interested in
We don't even know what this table is used for and how it is supposed
to behave.

The table we are actually interested in is nemeth.ctb. We can't test
it directly in Liblouis however because it is not part of
Liblouis (anymore). It is part of Liblouistdml. The solution is to
test nemethdefs.cti for now and add a note that we should change this
to nemeth.ctb (which requires that we move the table back to Liblouis,
see https://github.com/liblouis/liblouis/issues/815).

Also revert the changes in en-us-mathtext.ctb.